+++ tailwind.config.js
@@ -4,6 +4,9 @@ export default {
   content: [
     "./index.html",
     "./src/**/*.{js,ts,jsx,tsx,vue}",
+    "./cards/src/**/*.{svelte,js,ts}",
+    "./shared/src/**/*.{svelte,js,ts}",
+    "./cards-vue/src/**/*.{js,ts,vue}",
   ],
   theme: {
     extend: {},